+++ code/galpy/db_table_utility.py
@@ -247,24 +247,6 @@ class DefaultVariables:
                                        self.sequencing_center_contact_ID)
 
 
-class BaseCount:
-    def __init__(self, sequence):
-        self.sequence = sequence
-        self.A_count = len(re.findall("(?i)a", self.sequence))
-        self.T_count = len(re.findall("(?i)t", self.sequence))
-        self.G_count = len(re.findall("(?i)g", self.sequence))
-        self.C_count = len(re.findall("(?i)c", self.sequence))
-
-    def length(self):
-        return len(self.sequence)
-
-    def other_count(self):
-        count = len(self.sequence) - self.A_count - self.T_count - self.G_count - self.C_count
-        return count
-
-    def print_base_count(self):
-        other_count = self.other_count()
-        return '{}\t{}\t{}\t{}\t{}'.format(self.A_count, self.T_count, self.G_count, self.C_count, other_count)
 
 
 def upload_gal_table_data(db_config, upload_dir, logger):
